added mail and sms sender libraries

+++ notification/src/main/java/com/microservice/erp/domain/dto/NoticeDto.java
@@ -3,13 +3,11 @@ package com.microservice.erp.domain.dto;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
-
 @Setter
 @Getter
 public class NoticeDto {
     //region private variables
-    private BigInteger noticeConfigurationId;
+    private Long noticeConfigurationId;
     private String year;
     //endregion
 }
